Improved changelog for debian packages.

+++ package/make_signed_src_deb.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
 
+if [ -z "$1" ]
+  then
+    echo "Please specify an Ubuntu release (ie precise)"
+    exit 1
+fi
+
 SVNVER=`svn info ../ | grep Revision | awk '{ print $2 }'`
 echo $SVNVER
 
-BUILD_FOLDER=qstlink2-0.$SVNVER
+BUILD_FOLDER=qstlink2-0.$SVNVER~$1
+ORIG_FILE=$BUILD_FOLDER.orig.tar.gz
 
+rm -f $ORIG_FILE
 rm -rf $BUILD_FOLDER
 mkdir $BUILD_FOLDER
 
@@ -18,7 +26,9 @@ echo "" | dh_make -n --single -e fabien.poussin@gmail.com -c gpl3
 cp -v ../debian/* debian/
 rm -vf debian/*.ex debian/*.EX debian/ex.*
 
-debuild -j4 -b
+/usr/bin/python ../svn2debcl.py $1
+
+debuild -j4 -S -sa
 
 cd ..
-rm -rf $BUILD_FOLDER *.build *.changes
+rm -rf $BUILD_FOLDER
